gpu: nvgpu: clk: pass u32 for event number
Change the type of the event_number parameter of
nvgpu_clk_notification_queue_alloc() from size_t to u32 since it's
used everywhere as a u32.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2001229
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/clk/clk_arb.c
@@ -43,7 +43,7 @@
 
 int nvgpu_clk_notification_queue_alloc(struct gk20a *g,
 				struct nvgpu_clk_notification_queue *queue,
-				size_t events_number) {
+				u32 events_number) {
 	queue->notifications = nvgpu_kcalloc(g, events_number,
 		sizeof(struct nvgpu_clk_notification));
 	if (!queue->notifications)
